requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
cssselect2==0.4.1
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
django-bootstrap4==2.3.1
django-chartjs==2.2.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-session-csrf==0.7.1
django-smart-selects==1.5.9
django-stdimage==5.3.0
django-weasyprint==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
filelock==3.0.12
gunicorn==20.1.0
html5lib==1.1
importlib-metadata==2.1.1
joblib==1.0.1
nltk==3.6.2
permission==0.4.1
Pillow==8.1.2
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
Pyphen==0.10.0
pytz==2021.1
regex==2021.4.4
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
static3==0.7.0
textblob==0.15.3
tinycss2==1.1.0
toml==0.10.2
tqdm==4.60.0
utils==1.0.1
virtualenv==20.4.6
virtualenv-clone==0.5.4
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.2.0
zipp==3.4.0
